feat： 1、重新梳理 BaseApp、AppActivityManager。 2、使用 AppStateManager 替代 AppStateTracker
pref：
1、升级依赖库版本
2、对 base 包下的类，按功能重新分类。

+++ src/main/java/com/imyyq/mvvm/utils/BitOperationUtil.kt
@@ -23,5 +23,5 @@ fun Int.removeFlag(flag: Int) = this and flag.inv()
  * @return 返回 byte
  */
 fun Int.getByte(@IntRange(from = 0, to = 3) byteIndex: Int): Byte {
-    return (this as Int shr byteIndex * 8).toByte()
+    return (this shr byteIndex * 8).toByte()
 }
